owners: extract owners module
