+++ pom.xml
@@ -15,6 +15,7 @@
         <module>runner-quarkus</module>
         <module>api-tests</module>
         <module>war</module>
+        <module>vets</module>
     </modules>
 
     <groupId>net.kemitix.naolo</groupId>
