+++ vets/src/main/java/net/kemitix/naolo/vets/ListVets.java
@@ -19,10 +19,9 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package net.kemitix.naolo.core.vets;
+package net.kemitix.naolo.vets;
 
 import lombok.Getter;
-import net.kemitix.naolo.entities.Veterinarian;
 import net.kemitix.naolo.storage.EntityRepository;
 import net.kemitix.naolo.storage.ListEntityRequest;
 import net.kemitix.naolo.storage.ListEntityUseCase;
