+++ vets/src/main/java/net/kemitix/naolo/vets/VetSpecialisation.java
@@ -19,7 +19,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package net.kemitix.naolo.entities;
+package net.kemitix.naolo.vets;
 
 /**
  * Specialisations of Veterinarians.
