+++ vets/src/main/java/net/kemitix/naolo/vets/Veterinarian.java
@@ -1,6 +1,7 @@
-package net.kemitix.naolo.entities;
+package net.kemitix.naolo.vets;
 
 import lombok.*;
+import net.kemitix.naolo.entities.JPAActivityListener;
 import net.kemitix.naolo.storage.HasId;
 
 import javax.persistence.*;
