+++ vets/src/main/java/net/kemitix/naolo/vets/VeterinarianRepository.java
@@ -1,6 +1,6 @@
-package net.kemitix.naolo.storage;
+package net.kemitix.naolo.vets;
 
-import net.kemitix.naolo.entities.Veterinarian;
+import net.kemitix.naolo.storage.AbstractEntityRepository;
 
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
