+++ vets/src/test/java/net/kemitix/naolo/vets/GetVetTest.java
@@ -1,6 +1,5 @@
-package net.kemitix.naolo.core.vets;
+package net.kemitix.naolo.vets;
 
-import net.kemitix.naolo.entities.Veterinarian;
 import net.kemitix.naolo.storage.EntityRepository;
 import org.assertj.core.api.WithAssertions;
 import org.junit.jupiter.api.DisplayName;
