+++ vets/src/test/java/net/kemitix/naolo/vets/ListVetsTest.java
@@ -1,9 +1,7 @@
-package net.kemitix.naolo.core.vets;
+package net.kemitix.naolo.vets;
 
 import net.kemitix.naolo.core.StreamZipper;
 import net.kemitix.naolo.core.Tuple;
-import net.kemitix.naolo.entities.VetSpecialisation;
-import net.kemitix.naolo.entities.Veterinarian;
 import net.kemitix.naolo.storage.EntityRepository;
 import org.assertj.core.api.WithAssertions;
 import org.junit.jupiter.api.DisplayName;
