+++ vets/src/test/java/net/kemitix/naolo/vets/UpdateVetTest.java
@@ -1,7 +1,5 @@
-package net.kemitix.naolo.core.vets;
+package net.kemitix.naolo.vets;
 
-import net.kemitix.naolo.entities.VetSpecialisation;
-import net.kemitix.naolo.entities.Veterinarian;
 import net.kemitix.naolo.storage.EntityRepository;
 import org.assertj.core.api.WithAssertions;
 import org.junit.jupiter.api.DisplayName;
