+++ vets/src/test/resources/jqwik.properties
@@ -1,0 +1,1 @@
+
