+++ visits/src/test/java/net/kemitix/naolo/plugin/visits/AddVisitTest.java
@@ -1,8 +1,8 @@
 package net.kemitix.naolo.plugin.visits;
 
-import net.kemitix.naolo.entities.Veterinarian;
 import net.kemitix.naolo.pets.Pet;
 import net.kemitix.naolo.storage.EntityRepository;
+import net.kemitix.naolo.vets.Veterinarian;
 import org.assertj.core.api.WithAssertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
